der/requirements.txt
anthropic
langgraph
numpy
openai
protobuf
google-generativeai